theGreatPurge
Nah, despite the harrowing sound, not much changed here other than just cleaning up the repo of old folders and changing a few of the codes to work with the sceneManagement nonsense going on from an obscure Unity issue with projects that have strange characters. Note to self (and anyone that reads these), don't include anything weird or Unity will have a hissy fit, just use upper and lowercase letters for any projects for the best results.